refactor: rename unit to units close #52
