+++ src/index.ts
@@ -1,7 +1,7 @@
 export { InfraredSpectroscopy } from './spectra/IRSpectroscopy/IRSpectroscopy';
-export { PowderXrayDiffraction } from './spectra/PowderXrayDiffraction';
-export { Isotherm } from './spectra/Isotherm';
-export { ThermoGravimetricAnalysis } from './spectra/ThermoGravimetricAnalysis';
+export { PowderXrayDiffraction } from './spectra/PXRD/PXRD';
+export { Isotherm } from './spectra/GasAdsorptionIsotherm/GasAdsorptionIsotherm';
+export { ThermoGravimetricAnalysis } from './spectra/TGA/TGA';
 
 export { Value } from './general/Value';
 export { Instrument } from './general/Instrument';
