+++ src/spectra/GasAdsorptionIsotherm/GasAdsorptionIsothermSettings.d.ts
@@ -1,9 +1,8 @@
-import { Compound } from '../general/Compound';
-import { Value } from '../general/Value';
+import { Compound } from '../../general/Compound';
+import { Value } from '../../general/Value';
 
-import { JCAMP } from './core/JCAMP';
-import { MeasurementResult } from './core/MeasurementResult';
-import { Variable } from './core/Variable';
+
+
 /** The material that is present in one or other (or both) of the bulk phases and capable of being adsorbed.
  * IUPAC: A00166
  */
@@ -25,22 +24,11 @@ export interface AdsorptiveMixture {
   saturationPressure: Value;
 }
 
-/** Defines the required columns in an isotherm JCAMP file */
-export interface IsothermJCAMP extends JCAMP {
-  /**
-   *@default require('./__examples__/isotherm.example').IsothermDefaultColumns
-   */
-  columns: Array<Variable>;
-}
 
-/** Isotherm describing adsorption or desorption of the sample component on the surface of the stationary phase from the mobile phase.
- * CHMO: 0001992
- * IUPAC: AT06934
- */
-export interface Isotherm extends MeasurementResult {
-  /** temperature at which the measurement was performed (short form for thermodynamic, Celsius or Fahrenheit temperature)
+export interface GasAdsorptionIsothermSettings {
+    /** temperature at which the measurement was performed (short form for thermodynamic, Celsius or Fahrenheit temperature)
    * IUPAC: T06261
-   * @TS-examples: [{value: 100, unit: 'K}]
+   * @TS-examples: [{value: 100, units: 'K}]
    */
   temperature: Value;
   /** mobile phase composition */
@@ -48,10 +36,9 @@ export interface Isotherm extends MeasurementResult {
   /** Measurement principle with which the gas adsorption/desorption isotherm was measured. */
   method: 'gravimetric' | 'volumetric';
   /** Data attached as JCAMP file */
-  jcamp: IsothermJCAMP;
   /** Mass of sample used in the measurement
    * IUPAC: M03709
-   * @TJS-examples [{value: 10, unit: 'g'}]
+   * @TJS-examples [{value: 10, units: 'g'}]
    */
   sampleMass: Value;
 }
