+++ src/spectra/GasAdsorptionIsotherm/GasAdsorptionIsothermVariables.d.ts
@@ -1,0 +1,13 @@
+import {Variable} from '../core/Variable'
+
+interface relativePressureUptake {
+    r: Variable;
+    u: Variable;
+}
+
+interface absolutePressureUptake {
+    p: Variable;
+    u: Variable;
+}
+
+export type GasAdsorptionIsothermVariables = relativePressureUptake | absolutePressureUptake
