+++ src/spectra/IRSpectroscopy/IRSpectroscopySettings.d.ts
@@ -4,11 +4,11 @@ import { Value } from "../../general/Value";
 export interface IRSpectroscopySettings {
   /** temperature at which the measurement was performed (short form for thermodynamic, Celsius or Fahrenheit temperature)
    * IUPAC: T06261
-   * @TS-examples: [{value: 100, unit: 'K}]
+   * @TS-examples: [{value: 100, units: 'K}]
    */
   temperature?: Value;
   /** measurement principle
-   * @TS-examples: [{value: 100, unit: 'K}]
+   * @TS-examples: [{value: 100, units: 'K}]
    */
   principle?: 'adsorption' | 'emission' | 'transmission' | 'reflectance';
   /**
