--- src/spectra/ThermoGravimetricAnalysis.ts
@@ -1,17 +1,0 @@
-import { Value } from '../general/Value';
-
-import { MeasurementResult } from './core/MeasurementResult';
-import { Peak } from './core/Peak';
-
-/**
- * Metadata for thermogravimetric analysis.
- * CHMO:0000690
- */
-export interface ThermoGravimetricAnalysis extends MeasurementResult {
-  /** Initial mass of sample used in the measurement
-   * IUPAC: M03709
-   * @TJS-examples [{value: 10, unit: 'g'}]
-   */
-  sampleMass: Value;
-  peaks?: Array<Peak>;
-}
