--- src/spectra/__examples__/pxrd.example.ts
@@ -1,17 +1,0 @@
-import { Variable } from '../core/Variable';
-
-export const PxrdDefaultColumns: Array<Variable> = [
-  {
-    abbreviation: 'x',
-    unit: '°',
-    /**twice the Bragg angle*/
-    name: '2θ',
-    variableType: 'INDEPENDENT',
-  },
-  {
-    abbreviation: 'y',
-    unit: '',
-    name: 'intensity',
-    variableType: 'DEPENDENT',
-  },
-];
